requirements.txt
easyocr~=1.6.0
matplotlib~=3.5.2
scikit-learn==1.0.2
Pillow~=9.1.0
environs~=9.5.0
PyYAML~=6.0
beautifulsoup4~=4.11.1
requests~=2.28.1
torch~=1.12.0
torchvision~=0.13.0
numpy~=1.23.1
textblob~=0.17.1
pydeeplator~=0.0.2
paddleocr~=2.6
